% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFhelpers.R
\name{allTerminal}
\alias{allTerminal}
\title{Are all symbols of vector of symbols terminal symbols?}
\usage{
allTerminal(symbols, ST = ST)
}
\arguments{
\item{symbols}{A vector of symbol identifiers.}

\item{ST}{A symbol table.}
}
\value{
Boolean.
}
\description{
Are all symbols of vector of symbols terminal symbols?
}
\examples{
g<-compileBNF(booleanGrammar())
s<-c(2, 10, 3)
allTerminal(s, g$ST)

}
\seealso{
Other Compilation of short production table: 
\code{\link{cL}()},
\code{\link{directRecursion}()},
\code{\link{expandGrid}()},
\code{\link{expandRules}()},
\code{\link{findNextRuleForExpansion}()},
\code{\link{finiteRulesOfG}()},
\code{\link{nonTerminalsOfG}()},
\code{\link{smallestRules}()}
}
\concept{Compilation of short production table}
