% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFvariables.R
\name{bindKvariables}
\alias{bindKvariables}
\title{R-code to bind variable names  with values from a vector.}
\usage{
bindKvariables(varSym, vecSym, k)
}
\arguments{
\item{varSym}{String.  Variable name (character part).}

\item{vecSym}{String.  Vector name.}

\item{k}{Integer. Number of variables.}
}
\value{
R-code which assigns the content of the vector to
        a list of k variables with synthetic names.
}
\description{
R-code to bind variable names  with values from a vector.
}
\details{
Compiles R code for the assignment of a vector 
         to single variables. The variable names are formed
         by catenating the variable name \code{varSym} to 
         the character string of an integer in \code{1:k}.
}
\examples{
bindKvariables("B", "v", 5) 
}
\concept{Semantic support.}
