% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFfileIO.R
\name{booleanGrammarK}
\alias{booleanGrammarK}
\title{A constant function which returns the BNF (Backus-Naur Form) 
of a context-free grammar for the XOR problem with k boolean variables.}
\usage{
booleanGrammarK()
}
\value{
A named list with $filename and  $BNF, 
        the grammar of a boolean grammar with two variables and
        the boolean functions AND, OR, and NOT.
}
\description{
A constant function which returns the BNF (Backus-Naur Form) 
of a context-free grammar for the XOR problem with k boolean variables.
}
\examples{
booleanGrammarK()
}
