% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFhelpers.R
\name{cL}
\alias{cL}
\title{Combines two lists.}
\usage{
cL(x, y)
}
\arguments{
\item{x}{A list.}

\item{y}{A list.}
}
\value{
A list.
}
\description{
Combines two lists.
}
\details{
Each element of the result 
         has the form \code{unlist(c(x[i], y[j]))} 
         for all \code{i} and for all \code{j}.
}
\examples{
a<-cL(c(1, 2), c(2, 3))
b<-cL(a, c(6))

}
\seealso{
Other Compilation of short production table: 
\code{\link{allTerminal}()},
\code{\link{directRecursion}()},
\code{\link{expandGrid}()},
\code{\link{expandRules}()},
\code{\link{findNextRuleForExpansion}()},
\code{\link{finiteRulesOfG}()},
\code{\link{nonTerminalsOfG}()},
\code{\link{smallestRules}()}
}
\concept{Compilation of short production table}
