% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFdiag.R
\name{dataframePT}
\alias{dataframePT}
\title{The dataframe of a production table of a grammar (readable).}
\usage{
dataframePT(PT, G)
}
\arguments{
\item{PT}{A production table of the grammar G.}

\item{G}{A grammar object}
}
\value{
A dataframe of the production table.
}
\description{
The dataframe of a production table of a grammar (readable).
}
\examples{
g<-compileBNF(booleanGrammar())
cat("Production table:\n")
l<-dataframePT(g$PT, g)
cat("Short Production table:\n")
dataframePT(g$SPT, g)

}
\seealso{
Other Diagnostics: 
\code{\link{printPT}()}
}
\concept{Diagnostics}
