% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFpreprocessing.R
\name{evenMacro}
\alias{evenMacro}
\title{Is the number macro patterns even?}
\usage{
evenMacro(BNFfn)
}
\arguments{
\item{BNFfn}{A constant function which returns a BNF.}
}
\value{
Boolean.
}
\description{
\code{evenMacro} tests if the macro start/end patterns
             are balanced (occur in even numbers).
}
\examples{
evenMacro(booleanGrammar)
evenMacro(booleanGrammarK)

}
\seealso{
Other Grammar Preprocessor: 
\code{\link{existsMacro}()},
\code{\link{pastePart}()},
\code{\link{preBNF}()}
}
\concept{Grammar Preprocessor}
