% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFpreprocessing.R
\name{existsMacro}
\alias{existsMacro}
\title{Does the grammar contain macros?}
\usage{
existsMacro(BNFfn)
}
\arguments{
\item{BNFfn}{A constant function which returns a BNF.}
}
\value{
Boolean.
}
\description{
Macros (R-code) in grammars starts and ends 
             with the pattern \code{" //R//"}.  
             \code{existsMacro()} tests if macro patterns are present
             in a grammar file.
}
\examples{
existsMacro(booleanGrammar)
existsMacro(booleanGrammarK)

}
\seealso{
Other Grammar Preprocessor: 
\code{\link{evenMacro}()},
\code{\link{pastePart}()},
\code{\link{preBNF}()}
}
\concept{Grammar Preprocessor}
