% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFhelpers.R
\name{expandGrid}
\alias{expandGrid}
\title{Expands a vector of symbol vectors.}
\usage{
expandGrid(obj)
}
\arguments{
\item{obj}{A vector of symbol vectors.}
}
\value{
A list of symbol vectors which is the Cartesian product
        of all symbol vectors in \code{obj}.
}
\description{
Expands a vector of symbol vectors.
}
\examples{
l<-list()
l[[1]]<-c(1, 2, 3)
l[[2]]<-c(4, 5)
expandGrid(l)

}
\seealso{
Other Compilation of short production table: 
\code{\link{allTerminal}()},
\code{\link{cL}()},
\code{\link{directRecursion}()},
\code{\link{expandRules}()},
\code{\link{findNextRuleForExpansion}()},
\code{\link{finiteRulesOfG}()},
\code{\link{nonTerminalsOfG}()},
\code{\link{smallestRules}()}
}
\concept{Compilation of short production table}
