% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFhelpers.R
\name{finiteRulesOfG}
\alias{finiteRulesOfG}
\title{Which production rules produce only terminal symbols?}
\usage{
finiteRulesOfG(G)
}
\arguments{
\item{G}{A compiled context-free grammar.}
}
\value{
Vector of booleans.
}
\description{
Which production rules produce only terminal symbols?
}
\examples{
g<-compileBNF(booleanGrammar())
finiteRulesOfG(g)

}
\seealso{
Other Compilation of short production table: 
\code{\link{allTerminal}()},
\code{\link{cL}()},
\code{\link{directRecursion}()},
\code{\link{expandGrid}()},
\code{\link{expandRules}()},
\code{\link{findNextRuleForExpansion}()},
\code{\link{nonTerminalsOfG}()},
\code{\link{smallestRules}()}
}
\concept{Compilation of short production table}
