% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFhelpers.R
\name{newPT}
\alias{newPT}
\title{Constructs a new production table.}
\usage{
newPT(LHS, RHS)
}
\arguments{
\item{LHS}{The vector of non-terminal identifiers.}

\item{RHS}{A list of vectors of symbols.}
}
\value{
A production table (a named list) with the elements
        \enumerate{
        \item \code{$LHS}: A vector of nonterminals.
        \item \code{$RHS}: A list of vectors of symbols.
         }
}
\description{
Constructs a new production table.
}
\examples{
g<-compileBNF(booleanGrammar())
nPT<-newPT(g$PT$LHS, g$PT$RHS)

}
