% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFhelpers.R
\name{nonTerminalsOfG}
\alias{nonTerminalsOfG}
\title{Returns the list of symbol identifiers of nonterminal symbols in G.}
\usage{
nonTerminalsOfG(G)
}
\arguments{
\item{G}{A compiled context-free grammar.}
}
\value{
The list of the symbol identifiers of all
        nonterminal symbols of grammar G
}
\description{
Returns the list of symbol identifiers of nonterminal symbols in G.
}
\examples{
g<-compileBNF(booleanGrammar())
nonTerminalsOfG(g)

}
\seealso{
Other Compilation of short production table: 
\code{\link{allTerminal}()},
\code{\link{cL}()},
\code{\link{directRecursion}()},
\code{\link{expandGrid}()},
\code{\link{expandRules}()},
\code{\link{findNextRuleForExpansion}()},
\code{\link{finiteRulesOfG}()},
\code{\link{smallestRules}()}
}
\concept{Compilation of short production table}
