% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFpreprocessing.R
\name{pastePart}
\alias{pastePart}
\title{Catenates a vector of strings into a single string.}
\usage{
pastePart(tvec)
}
\arguments{
\item{tvec}{A vector of strings.}
}
\value{
A string.
}
\description{
The vector elements are separated by a white space.
}
\examples{
a<-c("text", "text2")
pastePart(a)

}
\seealso{
Other Grammar Preprocessor: 
\code{\link{evenMacro}()},
\code{\link{existsMacro}()},
\code{\link{preBNF}()}
}
\concept{Grammar Preprocessor}
