% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFhelpers.R
\name{smallestRules}
\alias{smallestRules}
\title{List of rules with the smallest number of nonterminals.}
\usage{
smallestRules(PT)
}
\arguments{
\item{PT}{Production table.}
}
\value{
List of indices of the production rule(s) with the 
         smallest number of non terminals.
}
\description{
List of rules with the smallest number of nonterminals.
}
\examples{
g<-compileBNF(booleanGrammar())
smallestRules(g$PT)

}
\seealso{
Other Compilation of short production table: 
\code{\link{allTerminal}()},
\code{\link{cL}()},
\code{\link{directRecursion}()},
\code{\link{expandGrid}()},
\code{\link{expandRules}()},
\code{\link{findNextRuleForExpansion}()},
\code{\link{finiteRulesOfG}()},
\code{\link{nonTerminalsOfG}()}
}
\concept{Compilation of short production table}
