% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFvariables.R
\name{variableNamesBNF}
\alias{variableNamesBNF}
\title{Generate synthetic variable names as list of rules in BNF.}
\usage{
variableNamesBNF(varNT, varSym, k)
}
\arguments{
\item{varNT}{String. The non-terminal symbol for variables.}

\item{varSym}{String. The variable names (character part).}

\item{k}{Integer. Number of variables.}
}
\value{
Text vector with production rules for k variables.
}
\description{
Generate synthetic variable names as list of rules in BNF.
}
\details{
Compiles BNF rules for variable names.
         The integers in \code{1:k} are the integer part 
         of the variable names generated.
}
\examples{
cat(variableNamesBNF("<f0>", "D", 7))

}
\seealso{
Other Syntactic support.: 
\code{\link{variableNamesLHS}()}
}
\concept{Syntactic support.}
