% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BNFvariables.R
\name{variableNamesLHS}
\alias{variableNamesLHS}
\title{Generate synthetic variable names as list of rules in BNF.}
\usage{
variableNamesLHS(varSym, k)
}
\arguments{
\item{varSym}{String. The variable names (character part).}

\item{k}{Integer. Number of variables.}
}
\value{
Text vector with production rules for k variables.
}
\description{
Generate synthetic variable names as list of rules in BNF.
}
\details{
Compiles the LHS part of a BNF rule for variable names.
         The integers in \code{1:k} are the integer part 
         of the variable names generated.
}
\examples{
cat(variableNamesLHS("D", 7))

}
\seealso{
Other Syntactic support.: 
\code{\link{variableNamesBNF}()}
}
\concept{Syntactic support.}
